+++ Assets/Scripts/Ammunition/Shells/ShellCharacteristic.cs
@@ -6,7 +6,7 @@ namespace Ammunition.Shells
     public class ShellCharacteristic
     {
         public int absolutDamage;
-        public int currentDamage;
+        public float currentDamage;
         public int force;
         public float lifeTime;
     }
