+++ Assets/Scripts/Interfaces/Player/IDamageable.cs
@@ -2,6 +2,6 @@ namespace Interfaces.Player
 {
     public interface IDamageable
     {
-        void DoDamage(int damageValue);
+        void DoDamage(float damageValue);
     }
 }
